+++ sdk/cognitiveservices/azure-mgmt-cognitiveservices/generated_samples/create_or_update_vm_managed_compute_deployment.py
@@ -48,6 +48,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-03-15-preview/CreateOrUpdateVmManagedComputeDeployment.json
+# x-ms-original-file: 2026-05-15-preview/CreateOrUpdateVmManagedComputeDeployment.json
 if __name__ == "__main__":
     main()