+++ sdk/netapp/azure-mgmt-netapp/generated_samples/subvolumes_list.py
@@ -40,6 +40,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2026-04-01/Subvolumes_List.json
+# x-ms-original-file: 2026-04-15-preview/Subvolumes_List.json
 if __name__ == "__main__":
     main()