+++ sdk/network/azure-mgmt-network/generated_samples/network_interface_effective_nsg_list.py
@@ -37,6 +37,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-05-01/NetworkInterfaceEffectiveNSGList.json
+# x-ms-original-file: 2025-07-01/NetworkInterfaceEffectiveNSGList.json
 if __name__ == "__main__":
     main()