+++ sdk/network/azure-mgmt-network/generated_samples/network_virtual_appliance_saa_sput.py
@@ -48,6 +48,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-05-01/NetworkVirtualApplianceSaaSPut.json
+# x-ms-original-file: 2025-07-01/NetworkVirtualApplianceSaaSPut.json
 if __name__ == "__main__":
     main()