+++ sdk/network/azure-mgmt-network/generated_samples/private_link_service_list_all.py
@@ -35,6 +35,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2025-05-01/PrivateLinkServiceListAll.json
+# x-ms-original-file: 2025-07-01/PrivateLinkServiceListAll.json
 if __name__ == "__main__":
     main()