+++ sdk/network/azure-mgmt-network/generated_samples/virtual_network_gateway_nat_rule_put.py
@@ -48,6 +48,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-05-01/VirtualNetworkGatewayNatRulePut.json
+# x-ms-original-file: 2025-07-01/VirtualNetworkGatewayNatRulePut.json
 if __name__ == "__main__":
     main()