+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_express_route_circuit_peerings_operations.py
@@ -25,7 +25,7 @@ class TestNetworkManagementExpressRouteCircuitPeeringsOperations(AzureMgmtRecord
             resource_group_name=resource_group.name,
             circuit_name="str",
             peering_name="str",
-            api_version="2025-05-01",
+            api_version="2025-07-01",
         )
 
         # please add some check logic here by yourself
@@ -127,7 +127,7 @@ class TestNetworkManagementExpressRouteCircuitPeeringsOperations(AzureMgmtRecord
                 },
                 "type": "str",
             },
-            api_version="2025-05-01",
+            api_version="2025-07-01",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -140,7 +140,7 @@ class TestNetworkManagementExpressRouteCircuitPeeringsOperations(AzureMgmtRecord
             resource_group_name=resource_group.name,
             circuit_name="str",
             peering_name="str",
-            api_version="2025-05-01",
+            api_version="2025-07-01",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -152,7 +152,7 @@ class TestNetworkManagementExpressRouteCircuitPeeringsOperations(AzureMgmtRecord
         response = self.client.express_route_circuit_peerings.list(
             resource_group_name=resource_group.name,
             circuit_name="str",
-            api_version="2025-05-01",
+            api_version="2025-07-01",
         )
         result = [r for r in response]
         # please add some check logic here by yourself